后端接口(除overview2未完成)全部完成

+++ src/main/java/com/hd/home_disabled/entity/Project.java
@@ -42,7 +42,7 @@ public class Project {
 
     private Integer personCountSum; //服务人数总数
     private Integer personTimeSum;  //服务总人次
-    private Integer totalTimeSum;   //服务总时长
+    private Float totalTimeSum;   //服务总时长
     private Float averageTime;      //平均服务时长：总时长/总人次
 
     //数据创建信息
