+++ src/main/java/com/hd/home_disabled/model/statistic/ProjectStatistic.java
@@ -15,6 +15,6 @@ public class ProjectStatistic {
     private String name;    //服务项目名称
     private Integer personCountSum; //服务人数总数
     private Integer personTimeSum;  //服务总人次
-    private Integer totalTimeSum;   //服务总时长（分钟）
+    private Float totalTimeSum;   //服务总时长（分钟）
     private Float averageTime;      //平均服务时长：总时长/总人次
 }
